requirements.txt
discord.py>=2.3.2
pytest>=8.0.0
click>=8.1.3
xlwings>=0.30.0
pdf2image>=1.0.6
python-dotenv>=1.0.0